app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
pandas==2.2.3
python-whois==0.9.5
httpx==0.28.1
beautifulsoup4==4.13.4
resend==2.7.0
google-genai==1.14.0
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
redis==5.3.0
celery==5.4.0
lxml==5.4.0
abuse-finder==0.1.1
pysocks==1.7.1
httpx[socks]==0.28.1
